move code for whatsnew to new file to not over-complicate the common preset. add scaffolding for save-from-controls
